refactor: change role management API responses
Refactors the role management API to use resource collections for more structured responses.

This change introduces `RoleCollection` and `RoleSimpleCollection` to format the API responses for listing, retrieving, updating, and deleting roles.  It also modifies the controller to return data in a consistent format that includes pagination information for the index route.