refactor: renames method for generating resource codes and create code generation traits.
Renames the `getGeneratedCode` method in the `CodeGeneration`
library to `getGeneratedResourceCode` for clarity.

This change improves the code's readability and accurately
reflects the method's purpose, which is specifically for
generating resource codes, distinguishing it from potential
future methods for generating other types of codes.

Updates all service classes and seeders that use the method.

Create `HasCodeGeneration` and `AutoCodeGeneration` traits to simplify creation of transactional or resource model column code